apiHH and RetrofitClient moved
