+++ app/src/main/java/ru/practicum/android/diploma/hhApi/impl/RetrofitClient.kt
@@ -1,11 +1,11 @@
-package ru.practicum.android.diploma.hhApi
+package ru.practicum.android.diploma.hhApi.impl
 
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import ru.practicum.android.diploma.filter.data.impl.dto.ApiArea
 import ru.practicum.android.diploma.filter.data.impl.dto.ApiCountry
-import ru.practicum.android.diploma.hhApi.impl.ApiHH
+import ru.practicum.android.diploma.hhApi.ApiHH
 
 class RetrofitClient {
     suspend fun loadData(): Response<List<ApiCountry>> {
